+++ Pandora/Paradigm/Algebraic/One.hs
@@ -1,3 +1,0 @@
-module Pandora.Paradigm.Algebraic.One where
-
-data One = One