+++ Pandora/Pattern/Operation/One.hs
@@ -1,0 +1,3 @@
+module Pandora.Pattern.Operation.One where
+
+data One = One