+++ Pandora/Pattern/Transformation.hs
@@ -12,6 +12,7 @@ class Component category t u where
 	component :: category (t a) (u a)
 
 type Semimonoidal_ category p q t = Component (Tensor p category q) t t
+-- TODO: Monoidal category p q t =
 type Liftable_ category u t = Component category u (t u)
 type Lowerable_ category t u = Component category (t u) u